"Modernize" `load_enrollments_report`
- Handle time formats with bad timezones.
- Add a kwarg for the `header`.
